requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Web Framework
Flask>=2.3.0
Werkzeug>=2.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Machine Learning utilities
scikit-learn>=1.2.0

# Optional: for advanced models
# timm>=0.9.0  # PyTorch Image Models
# segmentation-models-pytorch>=0.3.0  # Pre-trained segmentation models
